#27 rename: TokenDto 위치 이동
